Fix bug in faild to start skycoin app
The -gui-dir='.' in args that passed to skycoin node will be resolved to absolute path, while in electron, this abs path is actually the root path, eg: /, the skycoin will check if the dir of the path append with 'dev' is exist, if exist, if will be treated as web gui dir. The /dev path of course does exist, and will become as the web gui dir, so all requests will failed.