Rename variables in create_rawtx.go
Rename spendCoins to spendableCoins, unconfirmedOuts to uncfmIncomingOuts and unconfirmedBalance to uncfmIncoming.

We try to collect minimal spendable outputs to match the coins we need to send, the spendable outputs are all the confirmed outputs exclude the unconfirmed spending. If the spendable coins are sufficient, we return the collected outputs directly. otherwise we will check if there're unconfirmed incoming outputs, and returns different error message base on whether the unconfirmed incoming + spendable coins are sufficient.