Refactor the cipher.Encrypt and cipher.Decrypt
Remove the base64 encoding during the encryption and decryption

Change the data format from <length><hash><data> to <hash><length><data>, the previous format, if we decrypt the data with wrong password, the length could be any value, and we can't check if the password is valid before getting the hash and data throught the length. After we change to the second format, we can check if the hash is matched directly, and if it's not matched, it's wrong password, if matched, then we could read out length, and get the data.